requirements.txt
# Python dependencies for Supabase upsert script
# Using compatible versions that avoid dependency conflicts
supabase==2.8.1
python-dotenv>=1.0.0
httpx>=0.24,<0.26
numpy>=1.19.0,<2.0.0
websockets>=13.0
